requirements.txt
fastapi==0.59.0
uvicorn==0.11.7
#scikit-image==0.14.2
opencv-python==4.1.2.30
numpy==1.19.2
six==1.15.0
pillow==8.3.2
imutils
loguru==0.5.3
grpcio==1.32.0
pyyaml==6.0
tensorflow==2.4.0
tflite==2.4.0